+++ tests/UrlTest.php
@@ -73,15 +73,6 @@ class UrlTest extends Test
                     'buildRelative' => '',
                 ],
             ],
-            'protocol-relative host with auth' => [
-                '//user:pass@example.com',
-                [
-                    'getUser' => 'user',
-                    'getPassword' => 'pass',
-                    'getHost' => 'example.com',
-                    'buildRelative' => '',
-                ],
-            ],
             'host with protocol' => [
                 'http://example.com',
                 [
@@ -100,16 +91,6 @@ class UrlTest extends Test
                     'buildRelative' => '',
                 ],
             ],
-            'host with protocol and auth' => [
-                'http://user:pass@example.com',
-                [
-                    'getUser' => 'user',
-                    'getPassword' => 'pass',
-                    'getScheme' => 'http',
-                    'getHost' => 'example.com',
-                    'buildRelative' => '',
-                ],
-            ],
             'absolute url' => [
                 'http://www.example.com/foo/bar.html',
                 [
@@ -120,11 +101,9 @@ class UrlTest extends Test
                 ],
             ],
             'url with all components' => [
-                'https://user:pass@example.com:88/foo/bar.html?foo=bar&baz%5B0%5D=zero&baz%5B1%5D=one#test',
+                'https://example.com:88/foo/bar.html?foo=bar&baz%5B0%5D=zero&baz%5B1%5D=one#test',
                 [
                     'getScheme' => 'https',
-                    'getUser' => 'user',
-                    'getPassword' => 'pass',
                     'getHost' => 'example.com',
                     'getFullHost' => 'example.com:88',
                     'getPort' => 88,
@@ -187,35 +166,35 @@ class UrlTest extends Test
                     break;
 
                 case 4:
-                    $url->setUser('john.smith');
-                    $url->setPassword('123456');
-
-                    $expectedMethodResults['getUser'] = 'john.smith';
-                    $expectedMethodResults['getPassword'] = '123456';
-                    $expectedMethodResults['build'] = 'ftp://john.smith:123456@localhost:8080';
-                    break;
-
-                case 5:
                     $url->setPath('foo/bar');
 
                     $expectedMethodResults['getPath'] = 'foo/bar';
-                    $expectedMethodResults['build'] = 'ftp://john.smith:123456@localhost:8080/foo/bar';
+                    $expectedMethodResults['build'] = 'ftp://localhost:8080/foo/bar';
                     $expectedMethodResults['buildRelative'] = 'foo/bar';
+                    break;
+
+                case 5:
+                    $url->setPreferredFormat(Url::RELATIVE);
+
+                    $expectedMethodResults['getPreferredFormat'] = Url::RELATIVE;
+                    $expectedMethodResults['build'] = 'foo/bar';
                     break;
 
                 case 6:
                     $url->setQuery(['param' => 'value']);
+                    $url->setPreferredFormat(Url::ABSOLUTE);
 
                     $expectedMethodResults['getQuery'] = ['param' => 'value'];
-                    $expectedMethodResults['build'] = 'ftp://john.smith:123456@localhost:8080/foo/bar?param=value';
+                    $expectedMethodResults['build'] = 'ftp://localhost:8080/foo/bar?param=value';
                     $expectedMethodResults['buildRelative'] = 'foo/bar?param=value';
+                    unset($expectedMethodResults['getPreferredFormat']);
                     break;
 
                 case 7:
                     $url->set('lorem', ['ipsum', 'dolor']);
 
                     $expectedMethodResults['getQuery'] = ['param' => 'value', 'lorem' => ['ipsum', 'dolor']];
-                    $expectedMethodResults['build'] = 'ftp://john.smith:123456@localhost:8080/foo/bar'
+                    $expectedMethodResults['build'] = 'ftp://localhost:8080/foo/bar'
                         . '?param=value&lorem%5B0%5D=ipsum&lorem%5B1%5D=dolor';
                     $expectedMethodResults['buildRelative'] = 'foo/bar?param=value&lorem%5B0%5D=ipsum&lorem%5B1%5D=dolor';
                     break;
@@ -225,7 +204,7 @@ class UrlTest extends Test
                     $url->add(['param' => 'new-value']);
 
                     $expectedMethodResults['getQuery'] = ['param' => 'new-value'];
-                    $expectedMethodResults['build'] = 'ftp://john.smith:123456@localhost:8080/foo/bar?param=new-value';
+                    $expectedMethodResults['build'] = 'ftp://localhost:8080/foo/bar?param=new-value';
                     $expectedMethodResults['buildRelative'] = 'foo/bar?param=new-value';
                     break;
 
@@ -233,7 +212,7 @@ class UrlTest extends Test
                     $url->setFragment('test-fragment');
 
                     $expectedMethodResults['getFragment'] = 'test-fragment';
-                    $expectedMethodResults['build'] = 'ftp://john.smith:123456@localhost:8080/foo/bar?param=new-value#test-fragment';
+                    $expectedMethodResults['build'] = 'ftp://localhost:8080/foo/bar?param=new-value#test-fragment';
                     $expectedMethodResults['buildRelative'] = 'foo/bar?param=new-value#test-fragment';
                     break;
 
@@ -241,7 +220,7 @@ class UrlTest extends Test
                     $url->removeAll();
 
                     $expectedMethodResults['getQuery'] = [];
-                    $expectedMethodResults['build'] = 'ftp://john.smith:123456@localhost:8080/foo/bar#test-fragment';
+                    $expectedMethodResults['build'] = 'ftp://localhost:8080/foo/bar#test-fragment';
                     $expectedMethodResults['buildRelative'] = 'foo/bar#test-fragment';
                     break;
 
@@ -249,27 +228,18 @@ class UrlTest extends Test
                     $url->setFragment(null);
 
                     $expectedMethodResults['getFragment'] = null;
-                    $expectedMethodResults['build'] = 'ftp://john.smith:123456@localhost:8080/foo/bar';
+                    $expectedMethodResults['build'] = 'ftp://localhost:8080/foo/bar';
                     $expectedMethodResults['buildRelative'] = 'foo/bar';
                     break;
 
                 case 12:
-                    $url->setUser(null);
-                    $url->setPassword(null);
-
-                    $expectedMethodResults['getUser'] = null;
-                    $expectedMethodResults['getPassword'] = null;
-                    $expectedMethodResults['build'] = 'ftp://localhost:8080/foo/bar';
-                    break;
-
-                case 13:
                     $url->setScheme(null);
 
                     $expectedMethodResults['getScheme'] = null;
                     $expectedMethodResults['build'] = '//localhost:8080/foo/bar';
                     break;
 
-                case 14:
+                case 13:
                     $url->setPort(null);
 
                     $expectedMethodResults['getPort'] = null;
@@ -277,7 +247,7 @@ class UrlTest extends Test
                     $expectedMethodResults['build'] = '//localhost/foo/bar';
                     break;
 
-                case 15:
+                case 14:
                     $url->setPath('');
 
                     $expectedMethodResults['getPath'] = '';
@@ -285,7 +255,7 @@ class UrlTest extends Test
                     $expectedMethodResults['buildRelative'] = '';
                     break;
 
-                case 16:
+                case 15:
                     $url->setHost(null);
 
                     $expectedMethodResults['getHost'] = null;
@@ -324,6 +294,26 @@ class UrlTest extends Test
         $this->assertNull($url->get('null-param'));
     }
 
+    function testShouldSetDefaultPreferredFormatToAbsolute()
+    {
+        $this->assertSame(Url::ABSOLUTE, (new Url())->getPreferredFormat());
+        $this->assertSame(Url::ABSOLUTE, Url::parse('foo')->getPreferredFormat());
+    }
+
+    function testShouldSetPreferredFormatViaConstructor()
+    {
+        $url = new Url(null, null, null, '', [], null, Url::RELATIVE);
+
+        $this->assertSame(Url::RELATIVE, $url->getPreferredFormat());
+    }
+
+    function testShouldSetPreferredFormatViaParse()
+    {
+        $url = Url::parse('foo', Url::RELATIVE);
+
+        $this->assertSame(Url::RELATIVE, $url->getPreferredFormat());
+    }
+
     private function assertUrlMethodResults(Url $url, array $expectedMethodResults)
     {
         $expectedMethodResults += [
@@ -334,16 +324,13 @@ class UrlTest extends Test
             'getFullHost' => $expectedMethodResults['getHost'] ?? null,
             'getPort' => null,
             'hasPort' => isset($expectedMethodResults['getPort']),
-            'getUser' => null,
-            'hasUser' => isset($expectedMethodResults['getUser']),
-            'getPassword' => null,
-            'hasPassword' => isset($expectedMethodResults['getPassword']),
             'getPath' => '',
             'hasPath' => ($expectedMethodResults['getPath'] ?? '') !== '',
             'getQuery' => [],
             'hasQuery' => !empty($expectedMethodResults['getQuery']),
             'getFragment' => null,
             'hasFragment' => isset($expectedMethodResults['getFragment']),
+            'getPreferredFormat' => Url::ABSOLUTE,
         ];
 
         foreach ($expectedMethodResults as $method => $expectedValue) {
@@ -353,21 +340,5 @@ class UrlTest extends Test
                 sprintf('Expected Url::%s() to yield the expected value', $method)
             );
         }
-    }
-
-    private function setServerProperties(array $serverProperties = [])
-    {
-        $_SERVER = $serverProperties
-            + [
-                'HTTPS' => null,
-                'QUERY_STRING' => null,
-                'HTTP_HOST' => null,
-                'REQUEST_URI' => null,
-                'HTTP_X_REWRITE_URL' => null,
-                'HTTP_REQUEST_URI' => null,
-                'SCRIPT_NAME' => null,
-                'PHP_SELF' => null,
-            ]
-            + $_SERVER;
     }
 }